remove rounding on all analyses except DESeq

+++ functions/analysis/norm_graphs.py
@@ -250,7 +250,7 @@ def getNorm(request, RID, stopList, PID):
             myBiom['type'] = 'OTU table'
             myBiom['date'] = str(datetime.datetime.now())
             myBiom['matrix_type'] = 'dense'
-            myBiom['matrix_element_type'] = 'int'
+            myBiom['matrix_element_type'] = 'float'
             myBiom["shape"] = shape
             myBiom['rows'] = taxaList
             myBiom['columns'] = nameList
@@ -579,7 +579,7 @@ def normalizeUniv(df, taxaDict, mySet, meth, reads, metaDF, iters, Lambda, RID, 
     ser1 = countDF.groupby(field)[mySet].sum()
     ser2 = ser1.stack()
     abundDF = pd.Series.to_frame(ser2, name='abund')
-    abundDF['abund'] = abundDF['abund'].astype(int)
+    #abundDF['abund'] = abundDF['abund'].astype(int)    #JUMP
     abundDF.reset_index(drop=False, inplace=True)
 
     normDF = pd.merge(abundDF, namesDF, left_on='otuid', right_on='otuid', how='inner')
